Minor fixes and refactors
