+++ packages/cli/src/CommunityNodes/helpers.ts
@@ -4,7 +4,7 @@
 import { promisify } from 'util';
 import { exec } from 'child_process';
 import { access as fsAccess, mkdir as fsMkdir } from 'fs/promises';
-
+import { Script } from 'vm';
 import axios from 'axios';
 import { UserSettings } from 'n8n-core';
 import { LoggerProxy, PublicInstalledPackage } from 'n8n-workflow';
@@ -235,3 +235,9 @@ export const isClientError = (error: Error): boolean => {
 export function isNpmError(error: unknown): error is { code: number; stdout: string } {
 	return typeof error === 'object' && error !== null && 'code' in error && 'stdout' in error;
 }
+
+export const loadClassInIsolation = (filePath: string, className: string) => {
+	const script = new Script(`new (require('${filePath}').${className})()`);
+	// eslint-disable-next-line @typescript-eslint/no-unsafe-return
+	return script.runInNewContext({ require });
+};
